Move Index (and tests) to StandaloneRules
