+++ Assets/Scripts/StandaloneRules/Index.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a69fc86976268148bf400e0dd8720af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
