+++ Assets/Scripts/StandaloneRules/StandaloneRulesTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 49bd9b005c2a48740b9d9ed01b1d5cfb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
