+++ Assets/Scripts/StandaloneRules/StandaloneRulesTests/StandaloneRulesTests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 10099049963f22d4bb6d51c80f804011
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
